requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
scikit-surprise>=1.1.3
matplotlib>=3.6.0
jupyter>=1.0.0
